\name{massi.test.probes}
\alias{massi.test.probes}
\docType{data}
\title{
The massi test probes
}
\description{
A data.frame containing 56 probes which correspond to y chromosome genes
}
\usage{data(massi.test.probes)}
\format{data.frame
  \describe{A data frame with 56 observations on the following 0 variables.
  }
}

\examples{
data(massi.test.probes)
}
\keyword{datasets}
